Completed the basic requirements for questions.
Setup fetching questions from the Open Trivia Database's API. Created a few components (Answer, Question, Quiz) for questions. Made it so the question will display with four multiple choice answers, and when the user selects an answer, the correct state is set (e.g., the count of correct answers, the state for the current question, etc).